pyproject.toml

[project]
name = "issue-forge"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
  "requests",
  "jsonschema",
  "pyyaml",
  "yamllint",
  "pytest"
]
